Refactor all flash related methods into a Phoenix.Controller.Flash helper
